[draft] Preliminary support for building with Dune
Dear CompCert developers, this is a proof-of-concept highlighting a
full build of x86 using the Dune build system.

The PR is not meant to be merged yet, but as a demonstration and to
gather feedback. In particular, it has a few quirks that should be
solved soon Dune upstream:

- no proper configuration
- mixed Coq/ML dirs require a hack
- only works with Dune master

IMHO, I think it would make sense to eventually finish and merge this
PR, either as the main build system, or as secondary one, to be used
in Coq's Continous Integration system.

Dune support is very interesting for Coq developers as it allows for
composed, incremental builds with Coq itself; this means a large
speed-up in practice.